pyproject.toml
[tool.poetry]
name = "rsa"
version = "0.1.0"
description = ""
authors = ["kota <52403688+yashikota@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pycryptodome = "^3.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
